Massive changes, primarily to clean up type casting, also to convert away from use of select() system call in favor of ncurses equivalents.  "make depend.bsd" and "make depend.sys5" now exist in e20/Makefile; Ifdefs.bsd and Ifdefs.sys5 support this. There may be more work on the ncurses stuff.

+++ e20/term/standard.c
@@ -162,7 +162,7 @@ int *count;
     Block {
 	Reg1 int conv;
 	*count = nr;     /* number left over - still raw */
-	conv = ocp - lexp;
+	conv = (int) (ocp - lexp);
 	while (nr-- > 0)
 	    *ocp++ = *icp++;
 	return conv;
